+++ app/Http/Controllers/ProfileController.php
@@ -55,14 +55,17 @@ class ProfileController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int $profileId
+     * @param  int $id
      * @return \Illuminate\Http\JsonResponse
      */
-    public function show(int $profileId)
+    public function show(int $id)
     {
-        $profile = Auth('api')->id() == $profileId
-            ? $this->model->show($profileId)
-            : $this->model->showPublic($profileId);
+        if (!Auth::user()->email_verified_at && Auth('api')->id() != $id) {
+            return response()->json(config('messages.401'), 403);
+        }
+        $profile = Auth('api')->id() == $id
+            ? $this->model->show($id)
+            : $this->model->showPublic($id);
         return $profile ? response()->json($profile) : response('null');
     }