+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   User roles
+-   Restrict routes to users with email verified
 
 ## [0.2.0] - 2020-07-04